update to esp-develop-8.1.3

+++ hw/misc/esp32_ana.c
@@ -25,7 +25,7 @@ static uint64_t esp32_ana_read(void *opaque, hwaddr addr, unsigned int size)
           r=0xFFFFFFFF;
           break;
     }
-    if(DEBUG) printf("esp32_ana_read  0x%04lx= 0x%08x\n",addr,r);
+    if(DEBUG) printf("esp32_ana_read  0x%04lx= 0x%08x\n",(unsigned long) addr,r);
 
     return r;
 }
@@ -34,7 +34,7 @@ static void esp32_ana_write(void *opaque, hwaddr addr, uint64_t value,
                                  unsigned int size) {
     Esp32AnaState *s = ESP32_ANA(opaque);
 
-    if(DEBUG) printf("esp32_ana_write 0x%04lx= 0x%08lx\n",addr, value);
+    if(DEBUG) printf("esp32_ana_write 0x%04lx= 0x%08lx\n",(unsigned long) addr, (unsigned long) value);
 
     if(addr == 0xC4) {
         //printf("wifi channel:%x %x\n",(int)value, (int)~value);
